+++ composeApp/src/commonMain/kotlin/de/ashman/ontrack/features/tracking/TrackUiUtils.kt
@@ -183,14 +183,14 @@ fun TrackStatus.getStatusIcon(isFilled: Boolean = false): ImageVector {
 }
 
 @Composable
-fun getRatingLabel(rating: Int?, maxRating: Int): String {
+fun getRatingLabel(rating: Double?, maxRating: Int): String {
     val ratingLabel = when (rating) {
         null -> stringResource(Res.string.no_rating)
-        1 -> stringResource(Res.string.rating_one)
-        2 -> stringResource(Res.string.rating_two)
-        3 -> stringResource(Res.string.rating_three)
-        4 -> stringResource(Res.string.rating_four)
-        5 -> stringResource(Res.string.rating_five)
+        1.0 -> stringResource(Res.string.rating_one)
+        2.0 -> stringResource(Res.string.rating_two)
+        3.0 -> stringResource(Res.string.rating_three)
+        4.0 -> stringResource(Res.string.rating_four)
+        5.0 -> stringResource(Res.string.rating_five)
         else -> ""
     }
